Apparently I never sent these in. (#176)
This change is specifically for if a reminder is public to addon a message linking to where it will be sent too later.

Discussion seemed to lean towards the 2 newlines + "In channel:" however this is still up for debate as I only got 3 responses to this inquiry.

(cherry picked from commit 9fb5ef5c9bd173205dc393e0c06d2777ff5fa1ad)